Protect against invalid commands and improve debug
Under some circumstances, ReadController is called with an invalid
pif ram command (it is in fact a CIC challenge answer - See code for
details). When the plugin treats this request, it corrupts the challenge
result by setting an error flag. This makes the challenge fail and
emulation stops.

Protect against this by doing a simple check on the received values.